new solution

+++ FVCalc.java
@@ -4,7 +4,7 @@ public class FVCalc
 	public static void main(String[] args)
 	{
 		int currentValue = Integer.parseInt(args[0]);
-		int rate = Integer.parseInt(args[1]);
+		double rate = Double.parseDouble(args[1]);
 		int n = Integer.parseInt(args[2]);
